# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: CAOS API
  version: 0.1.0
  description: CAOS — Creative AI Operating System. API unificada do ecossistema CAOS (Shell, Nexus, Studio).
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /themes:
    get:
      operationId: listThemes
      tags: [themes]
      summary: List all theme items
      responses:
        "200":
          description: List of themes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Theme"
    post:
      operationId: createTheme
      tags: [themes]
      summary: Create a new theme item (manually or from design extraction)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ThemeInput"
      responses:
        "201":
          description: Created theme
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Theme"
  /themes/{id}:
    get:
      operationId: getTheme
      tags: [themes]
      summary: Get theme detail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Theme detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Theme"
    delete:
      operationId: deleteTheme
      tags: [themes]
      summary: Delete a theme
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /themes/{id}/apply:
    post:
      operationId: applyTheme
      tags: [themes]
      summary: Apply a theme to a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ThemeApplyRequest"
      responses:
        "200":
          description: Theme applied result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ThemeApplyResult"
  /themes/extract:
    post:
      operationId: extractTheme
      tags: [themes]
      summary: Extract a theme from image analysis (Design Extractor)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ThemeExtractRequest"
      responses:
        "200":
          description: Extracted theme
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Theme"

  # ─── AI AGENTS ───────────────────────────────────────────────────
...
  /skills:
    get:
      operationId: listSkills
      tags: [skills]
      summary: List all skill modules
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of skills
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Skill"
    post:
      operationId: createSkill
      tags: [skills]
      summary: Create a new skill module
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SkillInput"
      responses:
        "201":
          description: Created skill
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Skill"
  /skills/{id}:
    get:
      operationId: getSkill
      tags: [skills]
      summary: Get skill detail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Skill detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Skill"
    patch:
      operationId: updateSkill
      tags: [skills]
      summary: Update a skill
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SkillUpdate"
      responses:
        "200":
          description: Updated skill
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Skill"
    delete:
      operationId: deleteSkill
      tags: [skills]
      summary: Delete a skill
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /skills/{id}/plug:
    post:
      operationId: plugSkill
      tags: [skills]
      summary: Plug a skill into a project or agent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SkillPlugRequest"
      responses:
        "200":
          description: Plug result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SkillPlugResult"

  # ─── PROJECTS ────────────────────────────────────────────────────
...
components:
  schemas:
...
    Theme:
      type: object
      required: [id, name, rarity, colors, fonts, layoutStyle, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        rarity:
          type: string
          enum: [Common, Rare, Epic, Legendary]
        colors:
          type: array
          items:
            type: string
        fonts:
          type: array
          items:
            type: string
        layoutStyle:
          type: string
        spacing:
          type: ["string", "null"]
        borderRadius:
          type: ["string", "null"]
        previewUrl:
          type: ["string", "null"]
        sourceImageUrl:
          type: ["string", "null"]
        tags:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
    ThemeInput:
      type: object
      required: [name, colors, layoutStyle]
      properties:
        name:
          type: string
        rarity:
          type: string
          enum: [Common, Rare, Epic, Legendary]
        colors:
          type: array
          items:
            type: string
        fonts:
          type: array
          items:
            type: string
        layoutStyle:
          type: string
        spacing:
          type: string
        borderRadius:
          type: string
        previewUrl:
          type: string
        sourceImageUrl:
          type: string
        tags:
          type: array
          items:
            type: string
    ThemeApplyRequest:
      type: object
      required: [projectId]
      properties:
        projectId:
          type: integer
    ThemeApplyResult:
      type: object
      required: [success, message, projectId, themeId]
      properties:
        success:
          type: boolean
        message:
          type: string
        projectId:
          type: integer
        themeId:
          type: integer
    ThemeExtractRequest:
      type: object
      required: [imageUrl, name]
      properties:
        imageUrl:
          type: string
        name:
          type: string
        style:
          type: string

    # ─── AGENT ───────────────────────────────────────────────────
...
    Skill:
      type: object
      required: [id, name, category, description, rarity, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
        description:
          type: string
        rarity:
          type: string
          enum: [Common, Rare, Epic, Legendary]
        icon:
          type: ["string", "null"]
        behavior:
          type: ["string", "null"]
        pluggedCount:
          type: integer
        createdAt:
          type: string
          format: date-time
    SkillInput:
      type: object
      required: [name, category, description]
      properties:
        name:
          type: string
        category:
          type: string
        description:
          type: string
        rarity:
          type: string
          enum: [Common, Rare, Epic, Legendary]
        icon:
          type: string
        behavior:
          type: string
    SkillUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        category:
          type: string
        icon:
          type: string
        behavior:
          type: string
    SkillPlugRequest:
      type: object
      required: [targetType, targetId]
      properties:
        targetType:
          type: string
          enum: [agent, project]
        targetId:
          type: integer
    SkillPlugResult:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
        message:
          type: string

    # ─── PROJECT ─────────────────────────────────────────────────
...